M    ungrib/src/rd_grib1.F90 M    ungrib/src/gribcode.F90 M    util/src/plotfmt.F90
Well, I think I've fixed the NCEP2 problem once and for all.
The new code is in /mmmtmp/bresch/dave. Someone please check it in and let me know when it's in the repository.

It turns out the NCEP2 data are bad in 1997. It's just bad luck that
the guy from CU was using that year. (Of course it might be bad in earlier years, too). NCEP has since changed the NCEP2 header slightly. The repaired code works on all NCEP2 and other supported grib1 files. If you find a problem, please let me know. (The bottom line is, reorder_it is never called, so it can no longer cause problems).

I also have some enhancements to the plotfmt program.



git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@294 86b71a92-4018-0410-97f8-d555beccfc3a