NEW FEATURE:
Add a check in metgrid for missing values in output fields, and if such values
are found, stop. A value is counted as missing if the user has not explicitly
specified fill_missing=<some value>, and if the point is not in some region 
that is masked out (e.g., a water point for the ST000010 field). This should 
allow the user the flexibility to continue anyway for fields where missing
values can simply be filled in with a sane constant, or where the value is not
needed anyway.

M    metgrid/src/process_domain_module.F


git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@404 86b71a92-4018-0410-97f8-d555beccfc3a