Add support for NCEP godas files. g1print was extensively modified to allow easier addition of parameter tables. g1print will now decode NCEP table 129 and table 131 (NARR) files properly.
plotfmt was modified to allow plotting of subregions of the intermediate
files. I added it to plot portions of the GFS grid. It works for all
GFS files and the godas files. It is activated by setting parameters in
the code and recompiling. It's a "power user" feature and will be
undocumented. (It is very simple to use by anyone looking at
the source code).



git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@472 86b71a92-4018-0410-97f8-d555beccfc3a