Fix bug in sixteen_pt() and wt_sixteen_pt(). If any of the sixteen surrounding points is outside of the tile, don't just return a missing value; instead, try the next interpolation scheme in the sequence, which might need fewer points.
Thanks to Jeremy Krieger of ARSC for finding this.

NOTE: Maybe instead of trying the next interpolation scheme, we should average among only the 
points that are in the tile, and treat those outside the tile as missing. However, the current
fix is much simpler to implement, since we don't need extra logic to loop over a subset of 
the sixteen points which are in the tile.

M    geogrid/src/interp_module.F


git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@478 86b71a92-4018-0410-97f8-d555beccfc3a