As part of the port of the WRF model to Win64 for the WRF 3.2 release I developed a port of the WPS to Win64.
In addition to the Win port, it also has a couple changes to make it 
compatible with the WRF3.2 release. Specifically, 
module_driver_constants has to be listed explicitly in a few places 
because we've added a lot of ONLY clauses to the USE statements in WRF.

Most of the other changes are to handle the build under Cygwin using the 
PGI compilers.  The Win versions of these compilers have some issues 
with symbolic links and absolute UNIX path names that start with / 
(because that's an option string under Windows).  I moved a set of 
defines for WRF_LIB from the postamble to the preamble to allow the 
settings in arch/configure.defaults to override them and make 
corresponding changes to the Config.pl script. Only a couple changes to 
the source itself: there are also a couple instances of UNIX/Linux 
include files that aren't available under Windows.  I've substituted or 
just removed these inside #ifdef _WIN32 conditionals.



git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@534 86b71a92-4018-0410-97f8-d555beccfc3a