For some users with the latest version of GNU cpp, the WRF and the WPS codes fail to build due to the introduction of c-styled comments.  In Fortran code these break compilation, of course. The fix requires two steps:
1) Remove the -C option from the cpp command.  This now removes 
all comments.  This fixes the newly added c-styled comments that
cpp added that started with "/*" and ended with "*/".  
Unfortunately, the C++-styled comments extend to the end of a 
line after the sequence "//", which is a Fortran string
concatenator.  We only want the "/*" comments imapcted.

2) Tell CPP to ignore the new-fangled CPP comments with the 
-traditional option.

Only two files need to be modified for WPS.  The "-C" option is
removed for all CPP commands.  The "-traditional" option was
already in place for each of the CPP commands, so no action required,
but it is all about completeness.

M    geogrid/util/plotgrid/src/Makefile
M    arch/configure.defaults



git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@792 86b71a92-4018-0410-97f8-d555beccfc3a