Update the g2 and w3 libraries to the latest NCEP versions g2lib-2.5.0 (which was obtained from http://www.nco.ncep.noaa.gov/pmb/codes/nwprod/lib/sorc/g2 because the tar file on http://www.nco.ncep.noaa.gov/pmb/codes/GRIB2/ is messed up). and w3lib-2.0.2.tar.
Most changes are to add additional grids to the tables, and some changes
to encoders (not used by WPS). NCEP decided to remove spaces from the 5-character
variable labels, so users will see a difference in g2print (e.g. U GRD -> UGRD).

The biggest change was to comment out the "FIX THIS" print in g2/gf_free.f.
After 10 years of not being fixed this print statement is clearly not
effective...



git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@829 86b71a92-4018-0410-97f8-d555beccfc3a