Initial implementation of the 'optional' keyword in the GEOGRID.TBL file.
This optional keyword can take on one of two values: 'no' and 'yes'. 
The default is for an entry in the GEOGRID.TBL file to be non-optional (mandatory).
If the specified resolution of data for a mandatory GEOGRID.TBL entry is not found, 
geogrid will halt with an error message. At the end of processing for a domain,
information is printed for any optional entries for which data were not found.

It is possible to specify the optional keyword separately for different
priority levels of a field.

M    geogrid/src/source_data_module.F
M    geogrid/src/process_tile_module.F
M    geogrid/src/output_module.F
M    geogrid/src/geogrid.F
M    metgrid/src/storage_module.F


git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@883 86b71a92-4018-0410-97f8-d555beccfc3a