Add new optional keyword 'filename_digits' to index file
Until now, we had been able to get by with five digits in
the names of tiles of static data (e.g., 00001-01200.00001-01200);
however, new datasets available at resolutions of several tens
of meters and covering large geographical areas can have more
than 99,999 points in a dimension. In order to accommodate both
existing and new datasets, we add a new optional keyword 
'filename_digits' to the index files for datasets; the default
value is 5, so there is no need to update the index files of
existing geographical data. Currently, only values of 5 or 6
are supported.

The motivating use case is the NLCD2011 data, available for
the CONUS at 30m resolution; the total dataset size is 
161190 x 104424. We can now provide the complete dataset by
adding "filename_digits=6" in the index file for the prepared
data (and using six digits in the names of tiles).


M    source_data_module.F


git-svn-id: https://svn-wrf-wps.cgd.ucar.edu/trunk@909 86b71a92-4018-0410-97f8-d555beccfc3a