Move code for processing a single intermediate file to a separate function
In order to generalize the horizontal interpolation of input datasets in metgrid
to work with either intermediate files or with MPAS netCDF files, migrate code
that is specific to processing an intermediate file to its own subroutine,
process_intermediate_fields. The plan is to have a similar subroutine that
specifically processes MPAS netCDF files.