Handle new keyword 'mpas_name' in METGRID.TBL
In order to associate field names in MPAS output with the field name that would
be found in an intermediate file, add a new METGRID.TBL keyword 'mpas_name'
that gives the name of an MPAS field associated with an entry. Also, this commit
adds a routine for converting between an MPAS field name and whatever name
matches the METGRID.TBL entry associated with that mpas_name value.