Add routine mpas_output_stagger( ) to return output staggering for an MPAS field
This new routine, given the name of an MPAS field, searches the entries of
the METGRID.TBL file to find an entry with a matching name for 'mpas_name',
then returns the output staggering from that entry. If no match is found in
the METGRID.TBL file, a default staggering of M (mass grid) is returned.