Add routine derive_mpas_fields( ) to metgrid
The new routine derive_mpas_fields is called after missing levels are filled
in for fields in MPAS and before the check for missing mandatory fields. This
routine is intended to derive some mandatory fields from those that are given
in MPAS output: for example, we need to compute temperature (TT) given
the MPAS 'theta' and 'pressure' fields.