Add section to compile script to print code/compiler/system information (#12)
Added a few lines of code to the 'compile' script in order to print out
the version of the code. This is currently hard-wired as Version 3.9. We will
either need to come up with a more elegant method to have this automatically
update with each version, or simply remember to change it with each new release.
There will also be print-outs stating the compiler and system information.