Enable user to explicitly not use optional GEOGRID.TBL entries
Previously, for a GEOGRID.TBL entry that was marked as optional and
that did exist on the filesystem, if that entry did not have a default
interpolation option and no existing interpolation options were selected
by the user in the namelist.wps, geogrid would stop.

Now, it is possible to not select among any of the interpolation options
for an optional GEOGRID.TBL entry without a default interpolation option,
and to have the geogrid program continue.

Thanks to Brian Reen (US Army Research Lab) for providing these changes.