Add missing SAVE attribute to derived-type module variables in process_domain_module
Several derived-type module variables in the process_domain_module module
are of types with default initialization of their components. According to
the Fortran 2003 standard:

C1107 (R1104) If an object of a type for which component-initialization is specified (R444) appears
         in the specification-part of a module and does not have the ALLOCATABLE or POINTER
         attribute, the object shall have the SAVE attribute.

This commit simply adds the require SAVE attribute.

Thanks to Brian Reen (US Army Research Lab) for pointing out this issue.