Only convert TT from potential temperature to temperature for MPAS input (#28)
The routine derive_mpas_fields(...) is used to convert the TT field from
potential temperature to temperature using PRESSURE. Inside this routine,
however, there was previously no check to verify that the TT and PRESSURE
fields were from MPAS and not from another source where TT was already
temperature.

Now, we set the fg_source in the header of stored fields to 'MPAS' when
the data come from an MPAS dataset, and we only convert TT when the fg_source
is equal to 'MPAS'.

Also, the derive_mpas_fields(...) routine was previously enclosed inside
a test for (.not. do_const_processing); this is superfluous, since the code
above this call already returns if do_const_processing is true.