Fix compilation issue with 'new_plvl' in the &ungrib namelist record
Certain version of the ifort compiler (15.0.1 and 16.0.2) failed to compile
read_namelist.F because the assumed-shape array 'new_plvl' was part of
the &ungrib namelist record.

Following a discussion on stackoverflow
(http://stackoverflow.com/questions/30902646/using-allocatable-assumed-size-arrays-with-namelist-read-write
we add a local allocatable array to the &ungrib namelist record, then copy the
result into the assumed-shape dummy argument array after reading the namelist.