Fix issue with complex packing in newer NCEP NAM files (#32)
The packaging of NCEP NAM GRIB2 files was changed the week of 20 March 2017, and
the WPS system had troubles reading in the height field correctly. This merge includes
a bug fix from NCEP for their g2 library to properly decode GRIB2 files using complex packing.
Prior to 20 March 2017 complex packing was not used by model providers. The NAM was
the first NCEP model to use this packing.

Note: This same fix was made available on the WPS v3.8.1 "known problems" page.