Update util/src/Makefile so that module_map_utils.o is removed upon cleaning
Previously, when running './clean', the module_map_utils.o file in util/src
was not removed; this commit adds module_map_utils.o to the definition of OBJS,
which lists all object files to be removed when making the 'clean' target.

Note that module_map_utils.o is only created when explicitly building
the plotgrids utility.