Fix compilation errors in G2 library when using PGI compilers
Recent versions of the PGI compilers apparently still don't fully support
the use of all intrinsics in parameter statements, though this is part of
the F2003 standard.

In principle, the changes in this commit could reduce performance, since
the variables 'two23' and 'two126' are no longer parameters, but are set
in each subroutine call to rdieee and mkieee.