Fix rotation of input winds: Cassini to earth relative
Modify the existing met_to_map routine to use the correct rotation
angle, alpha. Previously, the rotation angle was computed with the lat,lon
for the d(lat)/d(y) and d(lon)/d(y) from the interpolated data on the
WRF domain. The correct lat and lon for these d/dy values is to use the
lat lon from the original input data.

The d(lat)/d(y) and d(lon)/d(y) computations use the existing latlon_to_ij
and the inverse ij_to_latlon in the module_map_utils.F file. This module is
now required to be "used" in rotate_winds_module.F, and the associated
Makefile for metgrid is also modified to include this new dependency.