Modify check for global grid for Gaussian and lat/lon grids in rd_grib1.F.
The high-resolution EC113.1 output revealed a problem with the testing
for a global grid. We test to see if the number_of_points times delta_lon
is close to 360 degrees and if it's 'close' then we assume it's a global
grid. When delta_lon is inaccurate (since it's only 3 digits), the
closeness can be more than 1 degree. So, increase the limit to 2. This
should not be a problem unless someone creates a non-global grid that
is 358 degrees wide.