Fix occasional NaNs in XLONG_* fields for polar stereographic projections
With some compilers, and with particular choices of projection parameters,
NaNs may appear in the XLONG_* fields produced by geogrid due to rounding
errors in the two terms used in the computation of the longitude.

Specifically, the conversion from (i,j) coordinates to (lat,lon) coordinates
for the polar projection uses the arccosine of the ratio of the x-coordinate
of the point to the radius from the pole of the projection to the point when
computing the longitude. It can sometimes happen that the radius is computed
as slightly less than the x-coordinate, causing the code to call the ACOS
function with an argument that is outside the range [-1,1].

The solution in this commit is to bound the argument to the ACOS function.