Fix inaccuracies in latitude computation for polar stereographic projections
The latitude field for polar stereographic projections is computed by taking
the arcsine of a value, and when computed in single precision, this can lead
to visible, discrete steps in the latitude field near +/-90 latitude.

To address this issue, we promote the arguments to the ASIN function to double-
precision. There is apparently no need to promote any other intermediate
terms to double-precision.

A simple test case involves a polar stereographic grid that contains either
pole, and that has a grid distance of <O(100 m).