Add 'lowres' resolutions for all fields in GEOGRID.TBL.ARW
This commit adds a resolution identified as "lowres" to all mandatory fields in
the GEOGRID.TBL.ARW file. The intent is to be able to easily select a suitably
low resolution for static fields when running the WPS, e.g., in a docker container.

Specifically, the resolutions for each field that will be selected by
geog_data_res = 'lowres' are:

 topography: 5-min GMTED2010 data (NEW DATASET)
 landuse: 5-min MODIS w/lakes (NEW DATASET)
 soiltemp: 1-deg (this was the only resolution for this dataset, anyway)
 soilctop: 5-min (this dataset already existed)
 soilcbot: 5-min (this dataset already existed)
 albedo12m: 0.05-deg MODIS-based albedo (this dataset already existed)
 greenfrac: 5-min MODIS FPAR-based (NEW DATASET)
 lai12m: 10-min MODIS LAI (this dataset already existed)
 snoalb: 0.05-deg MODIS-based (this dataset already existed)
 con/var/oa[1-4]/ol[1-4]: 1-deg (these datasets already existed)
 varsso: 10-min (this dataset already existed)