Combined Vtable for CFSR V1 and V2 Data (#10)
A vtable that can be used for both CFSR and CFSV2 has been created and named Vtable.CFSR. This replaces Vtable.CFSR2_web, Vtable.CFSR_press_pgbh06, and Vtable.CFSR_sfc_flx06 (which have been removed from the directory). Instructions are included in the Vtable comments to inform users of the proper way to handle the varying CFSR data types. For V1, ungrib must be run twice - once for the .5 deg 3-d pressure file and again for the Gaussian surface file. In v2, all fields are included in the 0.5 deg 3-d file (similar to GFS), and all fields are available in 6-h forecast intervals. However, if a user wants to use the 0.2 degree Gaussian v2 surface fields, they would need to split the Vtable into two pieces as it is now.

Tests have been performed using the new Vtable to ensure that it works properly.