Add a new namelist variable, pmin, to &ungrib to specify the minimum pressure level to process.
Some new NCEP output adds fields as high as 40 Pa. This is too high for typical
wrf applications. Because users will disagree about what is 'too high', they can
now specify their desired minimum pressure level. Setting pmin works for both
grib1 (EC) and grib2 files. Default is set to 100 Pa (1 mb or about 48 km).
For reference 1000 Pa is about 31 km, 40 Pa is 70 km, and 2000 Pa is 27 km.
It is recommended that pmin be equal to or smaller than p_top_requested in the
wrf namelist. Setting pmin=0. is the same as the current behavior which is to process
every pressure level (level code=100) in the grib file.

Using pmin has the added benefit of reducing wall clock time as well as the size
of the intermediate files since stratospheric levels can now be skipped.

Tested on GFS/FV3 and ERA-I files.