Add additional logic in arch/preamble for finding WRF source directory
The make logic in arch/preamble for setting the WRF_DIR variable to the location
of the compiled WRF source code previously looked either for a directory named
../WRF or a directory named ../WRFV3. In anticipation of users downloading
the WRF code as a GitHub archive file, which will unpack to a name like
WRF-4.0.1, we need more sophisticated logic for finding this directory.

This commit adds checks for the WRF code in ../WRF-4.0.1 and ../WRF-4.0.
In future we will likely need a more general solution (e.g., to handle the
eventual WRF 4.1 release), but as an immediate fix, checking for the WRF
4.0 or 4.0.1 releases should be sufficient.