Add additional logic in arch/preamble for finding WRF-4.0.2 source directory
Similar to what was done in c701b7e6, we need to update logic in arch/preamble
for setting the WRF_DIR variable to handle WRF source code that may be found
in a directory named WRF-4.0.2.

As mentioned in the commit message for c701b7e6, a more general solution to
finding the WRF source is needed to avoid growing complexity of the logic
in arch/preamble.