Remove -f90= and -cc= options from MPI wrappers
Depending on the MPI implementation, it appear that the mpif90 and mpicc
wrappers may or may not pass the -f90= and -cc= options directly to
the underlying compiler, where these options may generate either warnings
or errors.

The original intent of using, e.g., "mpif90 -f90=gfortran" was to permit
this use of a different Fortran compiler than the default selected by
the MPI wrapper. However, this appears to be not widely supported, and
it may be preferable to just remove these options and to assume that
the default compilers selected by the MPI wrappers are correct.