Let configure script attempt to locate compiled WRF code or use $WRF_DIR
Rather than continuing to extend an increasingly long set of if-tests in
the 'arch/preamble' code, which is restricted by what can be done in a makefile,
use more sophisticated logic in the 'configure' shell script to either use
the WRF build specified by the environment variable $WRF_DIR, or to try to
locate a WRF build in one of several expected locations.

Once a directory has been located (either by $WRF_DIR or by searching in
the 'configure' script), that directory is passed to the 'Config.pl' script
with a new switch, -wrfdir, where it is pre-processed in to the 'arch/preamble'
contents.