Change Vtable.GFS so the default sea level pressure is MSLET
Replaces PR #97 "Change Vtable.GFS so the default sea level pressure is MSLET"

NCEP has included the MSLET sea level pressure field in its pressure-level
grib files since May 10, 2011. This is in addition to the PRMSL sea level
pressure field (the previous default SLP in WPS). The PRMSL field is
smoothed/truncated. For the lower horizontal resolutions of GFS output in
years past this wasn't a significant issue. For the current higher
resolutions, the smoothing is very noticeable. (See attached figures).
SLP may be used to compute surface pressure in real. WRF tests with this
change show only minor differences. The main impact of using MSLET instead
of PRMSL is if metgrid output is used for plotting or verification.

Impact on the SLP field outside the areas of tropical storms:
Any system with a small-scale minimum or maximum will be smoothed away in
PRMSL. Practically, only cyclones (tropical or extratropical) have a
noticeable difference with the new field.

The smoothed field (top figure) is PRMSL which is the current default.
The bottom figure is MSLET the proposed default SLP field in WPS.

SMOOTHED SLP
![met2](https://user-images.githubusercontent.com/20980146/50380790-5d002c00-0630-11e9-8d01-1399cfbf661b.gif)

UN-SMOOTHED SLP
![met3](https://user-images.githubusercontent.com/20980146/50380791-625d7680-0630-11e9-920c-e95109850181.gif)