Add LOGSFP entry to METGRID.TBL.ARW
Model-level ECMWF datasets typically provide a logarithm of the surface
pressure, which may be used by the calc_ecmwf_p utility to produce a full 3-d
pressure field. In order to silence warnings from metgrid when the LOGSFP field
is found in intermediate files, this commit adds an entry to the METGRID.TBL.ARW
file for this field. Since the LOGSFP field is not used by WRF, the "output=no"
specification is also added to the LOGSFP entry.