Fix mismatch integer kind in IAND intrinsic calls in g2 library
The calls to the IAND intrinsic in the g2 library's ilog2_2 and ilog2_1
function used arguments of 'i' and 'i-1', where 'i' is of kind=2 or kind=1.
The XL compiler fails to compile this code with the complaint that
the arguments to IAND are of mismatched kind. It appears that 'i-1' is
being promoted to the default integer kind, unless we declare the constant
with _2 or _1.