Initialize arrays in calc_ecmwf_p to avoid potential error in GHT calculation
When the calc_ecmwf_p utility is provided with a subset of ECMWF model levels,
it may happen that the allocated 'tt' array contains non-zero values (because
it was not previously initialized). This leads to errors in the calculation of
the GHT field due to the use of uninitialized memory.

Thanks to Hugo Hartmann (GitHub: hugohartmann) for identifying this issue
and suggesting the fix in this commit.