Fix scaling of starting latitude in GRIB2 datasets on a Gaussian grid
Prior to this commit, the code in rd_grib2.F compared the starting latitude for
a field on a Gaussian grid with the scaling factor, and if the starting latitude
was larger than the scaling factor, the starting latitude was scaled. However,
this fails when the starting latitude is negative, in which case no scaling is
applied, and downstream code fails due to an invalid starting latitude in
the resulting intermediate file.

The fix in this commit is to simply compare the magnitude of the starting
latitude with the scaling factor, and if the former exceeds the latter, then
scaling of the starting latitude is applied.