Add missing category 21 for landmask calculation with modis_15s_lake land use
The 'landmask_water' entry in the GEOGRID.TBL.ARW file for the 'modis_15s_lake'
land use dataset previously only identified category 17 as a water category,
leading to an incorrect landmask field in the geogrid output.

This commit updates the 'landmask_water' entry to identify both category 17
and category 21 as water categories.

Thanks to GitHub user markelg for finding the issue addressed by this commit.