Update build_hdate.F
I am doing some paleo-downscaling with WRF, and have to ungrib a set of files that contain data at 3-digit years (e.g. year 450). Ungrib.exe was unable to extract data from these files, as

hdate 450-01-01_00:00:00 > hsave 0000-00-00_00:00:00

was not evaluated as true in one of the internal loops.
This was apparently caused by a formatting error when hdate was created.

I simply turned the year format from 'i4' to 'i4.4' on lines 26,30,34 and 38; which forces a leading zero (just as in other formatted time units). I did the same in 'geth_newdate.F', too.	

hdate 0450-01-01_00:00:00 > hsave 0000-00-00_00:00:00

The above is now .true., so ungrib.exe works for 3-digit years as well.