Fix URL for NARR static data grib1 data
The information for users in the NARR Vtable accidentally includes
an additional letter after "www". Once this change is made, sure
enough, we can download the NARR static data.

This is not a compiled file, so has no impact on the source build.

This part of the NARR Vtable is not processed by ungrib, it is purely
a comment.

Changes to be committed:
    modified:   ungrib/Variable_Tables/Vtable.NARR