Add --nowrf option to configure script, for configuring without WRF
Only three of the WPS components -- geogrid, metgrid, and int2nc -- require
the WRF I/O libraries. Even if none of these components was needed, it was
previously still necessary to either compile WRF or to make manual changes to
the configure script.

This commit introduces logic in the configure script to check for a command-
line argument, --nowrf, that, if present, disables checks in the configure
script for a compiled WRF directory. This enables, for example, the compilation
of the ungrib program without needing to compile the WRF model.