Add zLib, png, and JasPerLib sources for WPS distribution
This commit adds zlib-1.2.7, libpng-1.2.50 and jasper-1.900.1 source codes to
the ungrib directory under 'extern'. The purpose is to include and compile
these dependencies as part of the WPS, rather than requiring the user to
install them.

The extern directory contains a Makefile, which can configure, compiles,
installs and cleans each library, but currently, this Makefile is not apart of
the WPS/ungrib build system. The libraries are installed into the extern
directory.

The WPS `make` command to uses the `-r` make option, which ignores implicit
rules. This causes errors to appear when compiling zlib, as it would first try
and compile the example target, which builds against zlib itself, which has not
been built yet. Overriding MAKEFLAGS at the top of the extern makefile removes
the `-r` option, allowing zlib to compile correctly without producing errors.