Add option to configure to allow internal building of zlib, png and Jasper
Similar to the --nowrf option, a new option has been added to the configure
script, --build-grib2-libs, which builds the dependencies needed for ungrib2 in
the ungrib/extern directory. It then builds ungrib, by linking against these
libraries.

Specifying --build-grib2-libs when calling configure prompts the Config.pl
script to set the linking and include directories to the extern directory.  As
well, it sets the BUILD_GRIB2_LIBS variable in configure.wps to be 'true',
which prompts the ungrib Makefile to build the ungrib/extern directory.