Fix incorrect argument rank in calls to ext_*_put_dom_ti_* in output_module.F
In calls to ext_{int,ncd,gr1}_put_dom_ti_{integer,real} in output_module.F, the
'Data' dummy argument is an array, but a scalar actual argument was provided.
This error was picked up by the GNU Fortran 11.1.0 compiler:

  output_module.f90:1733:41:

   1733 |                                          var_value, &
        |                                         1
  ......
   1761 |                                          var_value, &
        |                                         2
  Error: Rank mismatch between actual argument at (1) and actual argument at (2) (rank-1 and scalar)
  output_module.f90:1678:41:

   1678 |                                          var_value, &
        |                                         1
  ......
   1706 |                                          var_value, &
        |                                         2
  Error: Rank mismatch between actual argument at (1) and actual argument at (2) (rank-1 and scalar)

The fix adopted by this commit is to simply make the scalar var_value into a
temporary size-1 array with (/var_value/).