Add python domain visualization script based on plotgrids_new.ncl
This utility script takes the original intent of plotgrids[_new].ncl and
rewrites it in python and adds multiple convenience features.

Extra info about domains is plotted next to each domain near their
respective ij start (SW corner in the projected coordinates).

An analogous perspective view can be plotted alongside the domain view.
This perspective view is taken as a straight down look on the Earth from
a geostationary orbit altitude looking directly at `ref_lat`/`ref_lon`
This may aid in assessing the "quality" of the projection with respect
to distortion and map scale factor.

To add in further diagnostics, a verification flag prompts the loading
of the respective geogrid .nc files for the provided namelist and plots
a field (default `HGT_M)` to the projections using the mass staggered
lat/lon coordinates. This can then be compared to the domain boxes to
ensure a match or lack thereof if something has gone amiss.

The following domain checks are performed to ensure good domain layout:
* nested domain size matching `parent_grid_ratio`
* nested domain boundary at least 5 parent cells away from parent boundary
* nested domain overlapping within the same parent

Rotated lat-lon projection is NOT SUPPORTED

Typical usage:
```
./util/plot_domains.py
```